+++ packages/react-drylus/src/layout/Grid.tsx
@@ -8,7 +8,7 @@ import React, { useEffect, useRef } from 'react';
 
 import { Size, Speed } from '../enums';
 import { Responsive, Style } from '../types';
-import { useResponsiveProps } from '../utils';
+import { checkComponentProps, useResponsiveProps } from '../utils';
 
 const styles = {
   root: (cols: number) => css`
@@ -217,17 +217,11 @@ export const Grid = ({ responsive, ...rest }: GridProps) => {
   const originOffset = useRef({ top: 0, left: 0 });
   const controls = useAnimation();
 
-  const invalidChildren = React.Children.map(children as any, (x) => x).some(
-    (child: React.ReactElement) =>
-      child != null && child.type !== GridItem && !child.type.toString().includes('fragment'),
-  );
-  if (invalidChildren) {
-    console.warn('Grid should only accept GridItem as children');
-  }
-
   useEffect(() => {
     controls.start('visible');
   }, [animated]);
+
+  checkComponentProps({ children }, { children: GridItem });
 
   return (
     <motion.div